# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: go-fiber-starter API
  description: Go + Fiber v3 + GORM 기반 REST API 보일러플레이트. 성공/실패 모두 공통 Envelope로 응답한다(/livez /readyz /metrics 등 운영 프로브 제외).
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/tasks:
    post:
      summary: Task 생성
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEnvelope'
        '400':
          description: 본문 파싱 실패(malformed request body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 필드 검증 실패(INVALID_REQUEST, details에 필드별 사유)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 전역 rate limit 초과(RATE_LIMITED, IP별 RATE_LIMIT_PER_MINUTE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      summary: Task 목록(offset 페이지네이션)
      operationId: listTasks
      tags: [tasks]
      parameters:
        - name: page
          in: query
          required: false
          description: 페이지 번호(1 이상, 기본 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: 페이지 크기(1~100, 기본 20, 100 초과분은 클램프)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 목록과 페이지 메타
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListEnvelope'
        '429':
          description: 전역 rate limit 초과(RATE_LIMITED, IP별 RATE_LIMIT_PER_MINUTE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 안정적인 에러 식별자. 예 NOT_FOUND, INVALID_REQUEST, TASK_NOT_FOUND, INVALID_CREDENTIALS, RATE_LIMITED
        message:
          type: string
        details:
          type: array
          description: 필드별 상세(주로 422 검증 실패). validator ValidationErrors의 Field/ActualTag에서 온다.
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
          description: 위반한 validate 태그(예 required, min, max)
    TaskEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/TaskResponse'
    TaskListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/Error'

    # === Task(task.dto) ===
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        due_date:
          type: string
          format: date-time
          nullable: true
...
    TaskResponse:
      type: object
      required: [id, title, done, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        done:
          type: boolean
        due_date:
          type: string
          format: date-time
          nullable: true
          description: 마감일. 미설정 시 필드가 생략된다(omitempty).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # === 페이지네이션(pagination.PageMeta) ===
    PageMeta:
      type: object
      required: [page, limit, total_count, total_pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_count:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64

    # === 인증(auth.handler) ===
...